Always display NA as unquoted in the Variables pane
