+++ crates/harp/src/vector/character_vector.rs
@@ -18,6 +18,7 @@ use libr::STRSXP;
 
 use crate::object::RObject;
 use crate::utils::r_str_to_owned_utf8_unchecked;
+use crate::vector::formatted_vector::FormattedVectorCharacterOptions;
 use crate::vector::Vector;
 
 #[harp_macros::vector]
@@ -80,8 +81,15 @@ impl Vector for CharacterVector {
         r_str_to_owned_utf8_unchecked(*x)
     }
 
-    fn format_one(&self, x: Self::Type) -> String {
-        x
+    fn format_one(
+        &self,
+        x: Self::Type,
+        options: Option<&FormattedVectorCharacterOptions>,
+    ) -> String {
+        match options {
+            Some(options) => self.format_with_string_options(x, options),
+            None => x,
+        }
     }
 }
