+++ crates/harp/src/vector/factor.rs
@@ -16,6 +16,7 @@ use libr::SEXP;
 
 use crate::object::RObject;
 use crate::r_symbol;
+use crate::vector::formatted_vector::FormattedVectorCharacterOptions;
 use crate::vector::CharacterVector;
 use crate::vector::Vector;
 
@@ -73,7 +74,11 @@ impl Vector for Factor {
         *x
     }
 
-    fn format_one(&self, x: Self::Type) -> String {
+    fn format_one(
+        &self,
+        x: Self::Type,
+        _option: Option<&FormattedVectorCharacterOptions>,
+    ) -> String {
         self.levels.get_unchecked((x - 1) as isize).unwrap()
     }
 }
