+++ crates/harp/src/vector/logical_vector.rs
@@ -14,6 +14,7 @@ use libr::LOGICAL_ELT;
 use libr::SEXP;
 
 use crate::object::RObject;
+use crate::vector::formatted_vector::FormattedVectorCharacterOptions;
 use crate::vector::Vector;
 
 #[harp_macros::vector]
@@ -68,7 +69,11 @@ impl Vector for LogicalVector {
         *x == 1
     }
 
-    fn format_one(&self, x: Self::Type) -> String {
+    fn format_one(
+        &self,
+        x: Self::Type,
+        _option: Option<&FormattedVectorCharacterOptions>,
+    ) -> String {
         if x {
             String::from("TRUE")
         } else {
