+++ crates/harp/src/vector/numeric_vector.rs
@@ -14,6 +14,7 @@ use libr::REAL_ELT;
 use libr::SEXP;
 
 use crate::object::RObject;
+use crate::vector::formatted_vector::FormattedVectorCharacterOptions;
 use crate::vector::Vector;
 
 #[harp_macros::vector]
@@ -68,7 +69,11 @@ impl Vector for NumericVector {
         *x
     }
 
-    fn format_one(&self, x: Self::Type) -> String {
+    fn format_one(
+        &self,
+        x: Self::Type,
+        _option: Option<&FormattedVectorCharacterOptions>,
+    ) -> String {
         x.to_string()
     }
 }
