+++ crates/harp/src/vector/raw_vector.rs
@@ -13,6 +13,7 @@ use libr::RAW_ELT;
 use libr::SEXP;
 
 use crate::object::RObject;
+use crate::vector::formatted_vector::FormattedVectorCharacterOptions;
 use crate::vector::Vector;
 
 #[harp_macros::vector]
@@ -67,7 +68,11 @@ impl Vector for RawVector {
         *x
     }
 
-    fn format_one(&self, x: Self::Type) -> String {
+    fn format_one(
+        &self,
+        x: Self::Type,
+        _option: Option<&FormattedVectorCharacterOptions>,
+    ) -> String {
         format!("{:02x}", x)
     }
 }
